Creates fake api.
